Move pytest_validation/ to validation/
