+++ src/software/ai/hl/stp/play/ball_placement/BUILD
@@ -40,7 +40,7 @@ py_test(
     tags = ["exclusive"],
     deps = [
         "//software:conftest",
-        "//software/simulated_tests/pytest_validations:all_validations",
+        "//software/simulated_tests/validation",
         requirement("pytest"),
     ],
 )
