+++ src/software/ai/hl/stp/play/crease_defense/BUILD
@@ -39,7 +39,7 @@ py_test(
     ],
     deps = [
         "//software:conftest",
-        "//software/simulated_tests/pytest_validations:all_validations",
+        "//software/simulated_tests/validation",
         requirement("pytest"),
     ],
 )
