+++ src/software/ai/hl/stp/play/enemy_ball_placement/BUILD
@@ -31,7 +31,7 @@ py_test(
     tags = ["exclusive"],
     deps = [
         "//software:conftest",
-        "//software/simulated_tests/pytest_validations:all_validations",
+        "//software/simulated_tests/validation",
         requirement("pytest"),
     ],
 )
