+++ src/software/ai/hl/stp/play/enemy_ball_placement/enemy_ball_placement_play_test.py
@@ -2,7 +2,7 @@ import pytest
 
 import software.python_bindings as tbots_cpp
 from proto.play_pb2 import Play, PlayName
-from software.simulated_tests.pytest_validations.robot_enters_placement_region import *
+from software.simulated_tests.validation.robot_enters_placement_region import *
 from software.simulated_tests.simulated_test_fixture import (
     pytest_main,
 )
