+++ src/software/ai/hl/stp/play/enemy_free_kick/BUILD
@@ -45,7 +45,7 @@ py_test(
     ],
     deps = [
         "//software:conftest",
-        "//software/simulated_tests/pytest_validations:all_validations",
+        "//software/simulated_tests/validation",
         requirement("pytest"),
     ],
 )
