+++ src/software/ai/hl/stp/play/free_kick/BUILD
@@ -54,7 +54,7 @@ py_test(
     tags = ["exclusive"],
     deps = [
         "//software:conftest",
-        "//software/simulated_tests/pytest_validations:all_validations",
+        "//software/simulated_tests/validation",
         requirement("pytest"),
     ],
 )
