+++ src/software/ai/hl/stp/play/halt_play/halt_play_test.py
@@ -1,5 +1,5 @@
 import software.python_bindings as tbots_cpp
-from software.simulated_tests.pytest_validations.robot_speed_threshold import *
+from software.simulated_tests.validation.robot_speed_threshold import *
 from proto.message_translation.tbots_protobuf import create_world_state
 from proto.ssl_gc_common_pb2 import Team
 from software.simulated_tests.simulated_test_fixture import (
