+++ src/software/ai/hl/stp/play/kickoff_play_test.py
@@ -2,15 +2,15 @@ import pytest
 
 import software.python_bindings as tbots_cpp
 from proto.play_pb2 import Play, PlayName
-from software.simulated_tests.pytest_validations.robot_enters_region import *
-from software.simulated_tests.pytest_validations.ball_enters_region import *
+from software.simulated_tests.validation.robot_enters_region import *
+from software.simulated_tests.validation.ball_enters_region import *
 from proto.import_all_protos import *
 from proto.message_translation.tbots_protobuf import create_world_state
 from proto.ssl_gc_common_pb2 import Team
 from software.simulated_tests.simulated_test_fixture import (
     pytest_main,
 )
-from software.simulated_tests.pytest_validations.or_validation import OrValidation
+from software.simulated_tests.validation.or_validation import OrValidation
 
 
 @pytest.mark.parametrize("is_friendly_test", [True, False])
