+++ src/software/ai/hl/stp/play/shoot_or_chip_play_test.py
@@ -3,7 +3,7 @@ import math
 
 import software.python_bindings as tbots_cpp
 from proto.play_pb2 import Play, PlayName
-from software.simulated_tests.pytest_validations.ball_enters_region import *
+from software.simulated_tests.validation.ball_enters_region import *
 from software.simulated_tests.simulated_test_fixture import (
     pytest_main,
 )
