+++ src/software/ai/hl/stp/tactic/crease_defender/BUILD
@@ -59,7 +59,7 @@ py_test(
     ],
     deps = [
         "//software:conftest",
-        "//software/simulated_tests/pytest_validations:all_validations",
+        "//software/simulated_tests/validation",
         requirement("pytest"),
     ],
 )
