+++ src/software/ai/hl/stp/tactic/crease_defender/crease_defender_tactic_test.py
@@ -8,14 +8,14 @@ from proto.import_all_protos import (
     MaxAllowedSpeedMode,
 )
 from proto.message_translation.tbots_protobuf import create_world_state
-from software.simulated_tests.pytest_validations.ball_is_off_ground import (
+from software.simulated_tests.validation.ball_is_off_ground import (
     BallIsAlwaysOnGround,
     BallIsEventuallyOffGround,
 )
-from software.simulated_tests.pytest_validations.excessive_dribbling import (
+from software.simulated_tests.validation.excessive_dribbling import (
     NeverExcessivelyDribbles,
 )
-from software.simulated_tests.pytest_validations.robot_enters_region import (
+from software.simulated_tests.validation.robot_enters_region import (
     RobotEventuallyEntersRegion,
     RobotNeverEntersRegion,
 )
