+++ src/software/ai/hl/stp/tactic/dribble/BUILD
@@ -74,7 +74,7 @@ py_test(
     ],
     deps = [
         "//software:conftest",
-        "//software/simulated_tests/pytest_validations:all_validations",
+        "//software/simulated_tests/validation",
         requirement("pytest"),
     ],
 )
