+++ src/software/ai/hl/stp/tactic/pass_defender/BUILD
@@ -45,7 +45,7 @@ py_test(
     ],
     deps = [
         "//software:conftest",
-        "//software/simulated_tests/pytest_validations:all_validations",
+        "//software/simulated_tests/validation",
         requirement("pytest"),
     ],
 )
