+++ src/software/ai/hl/stp/tactic/pass_defender/pass_defender_tactic_test.py
@@ -2,13 +2,13 @@ import pytest
 
 import software.python_bindings as tbots_cpp
 from proto.import_all_protos import PassDefenderTactic, BallStealMode
-from software.simulated_tests.pytest_validations.robot_enters_region import *
-from software.simulated_tests.pytest_validations.ball_enters_region import *
-from software.simulated_tests.pytest_validations.ball_moves_in_direction import *
-from software.simulated_tests.pytest_validations.friendly_has_ball_possession import *
-from software.simulated_tests.pytest_validations.ball_speed_threshold import *
-from software.simulated_tests.pytest_validations.robot_speed_threshold import *
-from software.simulated_tests.pytest_validations.excessive_dribbling import *
+from software.simulated_tests.validation.robot_enters_region import *
+from software.simulated_tests.validation.ball_enters_region import *
+from software.simulated_tests.validation.ball_moves_in_direction import *
+from software.simulated_tests.validation.friendly_has_ball_possession import *
+from software.simulated_tests.validation.ball_speed_threshold import *
+from software.simulated_tests.validation.robot_speed_threshold import *
+from software.simulated_tests.validation.excessive_dribbling import *
 from software.simulated_tests.simulated_test_fixture import (
     pytest_main,
 )
