+++ src/software/ai/navigator/trajectory/simulated_hrvo_test.py
@@ -2,7 +2,7 @@ import pytest
 from software.simulated_tests.simulated_test_fixture import (
     pytest_main,
 )
-from software.simulated_tests.pytest_validations.avoid_collisions import *
+from software.simulated_tests.validation.avoid_collisions import *
 import software.python_bindings as tbots
 from software.py_constants import *
 from proto.message_translation.tbots_protobuf import create_world_state
@@ -10,7 +10,7 @@ import math
 from proto.import_all_protos import *
 from proto.ssl_gc_common_pb2 import Team
 from software.simulated_tests.simulated_test_fixture import SimulatedTestRunner
-from software.simulated_tests.pytest_validations.validation import (
+from software.simulated_tests.validation.validation import (
     create_validation_types,
     create_validation_geometry,
 )
