+++ src/software/field_tests/passing_field_test.py
@@ -1,7 +1,7 @@
 import software.python_bindings as tbots_cpp
 from proto.import_all_protos import *
 from software.field_tests.field_test_fixture import *
-from software.simulated_tests.pytest_validations.friendly_receives_ball_slow import (
+from software.simulated_tests.validation.friendly_receives_ball_slow import (
     FriendlyAlwaysReceivesBallSlow,
 )
 from software.simulated_tests.simulated_test_fixture import (
