+++ src/software/simulated_tests/simulated_test_ball_model.py
@@ -1,14 +1,14 @@
 import pytest
 
 import software.python_bindings as tbots_cpp
-from software.simulated_tests.pytest_validations.robot_enters_region import *
-from software.simulated_tests.pytest_validations.ball_enters_region import *
-from software.simulated_tests.pytest_validations.ball_moves_in_direction import *
-from software.simulated_tests.pytest_validations.friendly_has_ball_possession import *
-from software.simulated_tests.pytest_validations.ball_speed_threshold import *
-from software.simulated_tests.pytest_validations.robot_speed_threshold import *
-from software.simulated_tests.pytest_validations.ball_stops_in_region import *
-from software.simulated_tests.pytest_validations.excessive_dribbling import *
+from software.simulated_tests.validation.robot_enters_region import *
+from software.simulated_tests.validation.ball_enters_region import *
+from software.simulated_tests.validation.ball_moves_in_direction import *
+from software.simulated_tests.validation.friendly_has_ball_possession import *
+from software.simulated_tests.validation.ball_speed_threshold import *
+from software.simulated_tests.validation.robot_speed_threshold import *
+from software.simulated_tests.validation.ball_stops_in_region import *
+from software.simulated_tests.validation.excessive_dribbling import *
 from proto.message_translation.tbots_protobuf import create_world_state
 from software.simulated_tests.simulated_test_fixture import (
     pytest_main,
