+++ src/software/simulated_tests/simulated_test_fixture.py
@@ -8,7 +8,7 @@ import os
 import pytest
 from proto.import_all_protos import *
 
-from software.simulated_tests.pytest_validations import validation
+from software.simulated_tests.validation import validation
 from software.simulated_tests.tbots_test_runner import TbotsTestRunner
 from software.thunderscope.thunderscope import Thunderscope
 from software.thunderscope.proto_unix_io import ProtoUnixIO
