+++ src/software/simulated_tests/validation/avoid_collisions.py
@@ -1,7 +1,7 @@
 from software.py_constants import *
 import software.python_bindings as tbots
 from proto.import_all_protos import *
-from software.simulated_tests.pytest_validations.validation import (
+from software.simulated_tests.validation.validation import (
     Validation,
     create_validation_types,
     create_validation_geometry,
