+++ src/software/simulated_tests/validation/ball_moves_in_direction.py
@@ -1,7 +1,7 @@
 import software.python_bindings as tbots
 from proto.import_all_protos import *
 
-from software.simulated_tests.pytest_validations.validation import (
+from software.simulated_tests.validation.validation import (
     Validation,
     create_validation_geometry,
     create_validation_types,
