+++ src/software/simulated_tests/validation/or_validation.py
@@ -1,5 +1,5 @@
 from proto.validation_pb2 import *
-from software.simulated_tests.pytest_validations.validation import (
+from software.simulated_tests.validation.validation import (
     Validation,
 )
 from typing import override
