+++ src/software/simulated_tests/validation/robot_at_orientation.py
@@ -1,7 +1,7 @@
 import software.python_bindings as tbots_cpp
 from proto.import_all_protos import *
 
-from software.simulated_tests.pytest_validations.validation import (
+from software.simulated_tests.validation.validation import (
     Validation,
     create_validation_geometry,
     create_validation_types,
