+++ src/software/simulated_tests/validation/robot_enters_region_and_stops.py
@@ -2,11 +2,11 @@ import math
 from software.py_constants import *
 from proto.import_all_protos import *
 
-from software.simulated_tests.pytest_validations.validation import (
+from software.simulated_tests.validation.validation import (
     create_validation_types,
     create_validation_geometry,
 )
-from software.simulated_tests.pytest_validations.robot_enters_region import (
+from software.simulated_tests.validation.robot_enters_region import (
     RobotEntersRegion,
 )
 from typing import override
