+++ src/software/simulated_tests/validation/robot_speed_threshold.py
@@ -1,11 +1,11 @@
 import software.python_bindings as tbots_cpp
 from proto.import_all_protos import *
 from software.py_constants import *
-from software.simulated_tests.pytest_validations.speed_threshold_helpers import *
+from software.simulated_tests.validation.speed_threshold_helpers import *
 from typing import override
 
 
-from software.simulated_tests.pytest_validations.validation import (
+from software.simulated_tests.validation.validation import (
     Validation,
     create_validation_geometry,
     create_validation_types,
